chores: #32 fix the code layout
